Trait rusoto_iot1click_projects::Iot1ClickProjects

pub trait Iot1ClickProjects {
    fn associate_device_with_placement(
        &self, 
        input: AssociateDeviceWithPlacementRequest
    ) -> RusotoFuture<AssociateDeviceWithPlacementResponse, AssociateDeviceWithPlacementError>;

    fn create_placement(
        &self, 
        input: CreatePlacementRequest
    ) -> RusotoFuture<CreatePlacementResponse, CreatePlacementError>;

    fn create_project(
        &self, 
        input: CreateProjectRequest
    ) -> RusotoFuture<CreateProjectResponse, CreateProjectError>;

    fn delete_placement(
        &self, 
        input: DeletePlacementRequest
    ) -> RusotoFuture<DeletePlacementResponse, DeletePlacementError>;

    fn delete_project(
        &self, 
        input: DeleteProjectRequest
    ) -> RusotoFuture<DeleteProjectResponse, DeleteProjectError>;

    fn describe_placement(
        &self, 
        input: DescribePlacementRequest
    ) -> RusotoFuture<DescribePlacementResponse, DescribePlacementError>;

    fn describe_project(
        &self, 
        input: DescribeProjectRequest
    ) -> RusotoFuture<DescribeProjectResponse, DescribeProjectError>;

    fn disassociate_device_from_placement(
        &self, 
        input: DisassociateDeviceFromPlacementRequest
    ) -> RusotoFuture<DisassociateDeviceFromPlacementResponse, DisassociateDeviceFromPlacementError>;

    fn get_devices_in_placement(
        &self, 
        input: GetDevicesInPlacementRequest
    ) -> RusotoFuture<GetDevicesInPlacementResponse, GetDevicesInPlacementError>;

    fn list_placements(
        &self, 
        input: ListPlacementsRequest
    ) -> RusotoFuture<ListPlacementsResponse, ListPlacementsError>;

    fn list_projects(
        &self, 
        input: ListProjectsRequest
    ) -> RusotoFuture<ListProjectsResponse, ListProjectsError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_placement(
        &self, 
        input: UpdatePlacementRequest
    ) -> RusotoFuture<UpdatePlacementResponse, UpdatePlacementError>;

    fn update_project(
        &self, 
        input: UpdateProjectRequest
    ) -> RusotoFuture<UpdateProjectResponse, UpdateProjectError>;
}

Trait representing the capabilities of the AWS IoT 1-Click Projects API. AWS IoT 1-Click Projects clients implement this trait.

Required methods

fn associate_device_with_placement(
&self,
input: AssociateDeviceWithPlacementRequest
) -> RusotoFuture<AssociateDeviceWithPlacementResponse, AssociateDeviceWithPlacementError>

Associates a physical device with a placement.

fn create_placement(
&self,
input: CreatePlacementRequest
) -> RusotoFuture<CreatePlacementResponse, CreatePlacementError>

Creates an empty placement.

fn create_project(
&self,
input: CreateProjectRequest
) -> RusotoFuture<CreateProjectResponse, CreateProjectError>

Creates an empty project with a placement template. A project contains zero or more placements that adhere to the placement template defined in the project.

fn delete_placement(
&self,
input: DeletePlacementRequest
) -> RusotoFuture<DeletePlacementResponse, DeletePlacementError>

Deletes a placement. To delete a placement, it must not have any devices associated with it.

When you delete a placement, all associated data becomes irretrievable.

fn delete_project(
&self,
input: DeleteProjectRequest
) -> RusotoFuture<DeleteProjectResponse, DeleteProjectError>

Deletes a project. To delete a project, it must not have any placements associated with it.

When you delete a project, all associated data becomes irretrievable.

fn describe_placement(
&self,
input: DescribePlacementRequest
) -> RusotoFuture<DescribePlacementResponse, DescribePlacementError>

Describes a placement in a project.

fn describe_project(
&self,
input: DescribeProjectRequest
) -> RusotoFuture<DescribeProjectResponse, DescribeProjectError>

Returns an object describing a project.

fn disassociate_device_from_placement(
&self,
input: DisassociateDeviceFromPlacementRequest
) -> RusotoFuture<DisassociateDeviceFromPlacementResponse, DisassociateDeviceFromPlacementError>

Removes a physical device from a placement.

fn get_devices_in_placement(
&self,
input: GetDevicesInPlacementRequest
) -> RusotoFuture<GetDevicesInPlacementResponse, GetDevicesInPlacementError>

Returns an object enumerating the devices in a placement.

fn list_placements(
&self,
input: ListPlacementsRequest
) -> RusotoFuture<ListPlacementsResponse, ListPlacementsError>

Lists the placement(s) of a project.

fn list_projects(
&self,
input: ListProjectsRequest
) -> RusotoFuture<ListProjectsResponse, ListProjectsError>

Lists the AWS IoT 1-Click project(s) associated with your AWS account and region.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Lists the tags (metadata key/value pairs) which you have assigned to the resource.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Creates or modifies tags for a resource. Tags are key/value pairs (metadata) that can be used to manage a resource. For more information, see AWS Tagging Strategies.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Removes one or more tags (metadata key/value pairs) from a resource.

fn update_placement(
&self,
input: UpdatePlacementRequest
) -> RusotoFuture<UpdatePlacementResponse, UpdatePlacementError>

Updates a placement with the given attributes. To clear an attribute, pass an empty value (i.e., "").

fn update_project(
&self,
input: UpdateProjectRequest
) -> RusotoFuture<UpdateProjectResponse, UpdateProjectError>

Updates a project associated with your AWS account and region. With the exception of device template names, you can pass just the values that need to be updated because the update request will change only the values that are provided. To clear a value, pass the empty string (i.e., "").

Loading content...