Crate rusoto_iot1click_projects

The AWS IoT 1-Click Projects API Reference

If you're using the service, you're probably looking for Iot1ClickProjectsClient and Iot1ClickProjects.

Structs

AssociateDeviceWithPlacementRequest

AssociateDeviceWithPlacementResponse

CreatePlacementRequest

CreatePlacementResponse

CreateProjectRequest

CreateProjectResponse

DeletePlacementRequest

DeletePlacementResponse

DeleteProjectRequest

DeleteProjectResponse

DescribePlacementRequest

DescribePlacementResponse

DescribeProjectRequest

DescribeProjectResponse

DeviceTemplate
An object representing a device for a placement template (see PlacementTemplate).

DisassociateDeviceFromPlacementRequest

DisassociateDeviceFromPlacementResponse

GetDevicesInPlacementRequest

GetDevicesInPlacementResponse

Iot1ClickProjectsClient
A client for the AWS IoT 1-Click Projects API.

ListPlacementsRequest

ListPlacementsResponse

ListProjectsRequest

ListProjectsResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

PlacementDescription
An object describing a project's placement.

PlacementSummary
An object providing summary information for a particular placement.

PlacementTemplate
An object defining the template for a placement.

ProjectDescription
An object providing detailed information for a particular project associated with an AWS account and region.

ProjectSummary
An object providing summary information for a particular project for an associated AWS account and region.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdatePlacementRequest

UpdatePlacementResponse

UpdateProjectRequest

UpdateProjectResponse

Enums

AssociateDeviceWithPlacementError
Errors returned by AssociateDeviceWithPlacement

CreatePlacementError
Errors returned by CreatePlacement

CreateProjectError
Errors returned by CreateProject

DeletePlacementError
Errors returned by DeletePlacement

DeleteProjectError
Errors returned by DeleteProject

DescribePlacementError
Errors returned by DescribePlacement

DescribeProjectError
Errors returned by DescribeProject

DisassociateDeviceFromPlacementError
Errors returned by DisassociateDeviceFromPlacement

GetDevicesInPlacementError
Errors returned by GetDevicesInPlacement

ListPlacementsError
Errors returned by ListPlacements

ListProjectsError
Errors returned by ListProjects

ListTagsForResourceError
Errors returned by ListTagsForResource

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdatePlacementError
Errors returned by UpdatePlacement

UpdateProjectError
Errors returned by UpdateProject

Traits

Iot1ClickProjects
Trait representing the capabilities of the AWS IoT 1-Click Projects API. AWS IoT 1-Click Projects clients implement this trait.