Struct rusoto_iot1click_projects::Iot1ClickProjectsClient

pub struct Iot1ClickProjectsClient { /* fields omitted */ }

A client for the AWS IoT 1-Click Projects API.

Methods

impl Iot1ClickProjectsClient

pub fn new(region: Region) -> Iot1ClickProjectsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> Iot1ClickProjectsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,